requirements.txt
django
djangorestframework
psycopg2-binary
django-jazzmin
django-health-check
psutil